refactor(contract): move contract-update proposals into api/update.rs
