+++ crates/contract/src/api.rs
@@ -11,3 +11,4 @@ pub mod sign;
 #[cfg(not(target_arch = "wasm32"))]
 #[cfg(test)]
 pub mod test_utils;
+pub mod update;
